Creates default vars for custom elk role
Rather empty at the moment, but let's keep things moving along and
circle back on the customization. How often are you going to want to
downgrade to Kibana 3 anyway?

Creates default var for Kibana version;
4.3.1 is the newest, automatically interpolated the var everywhere for
download and skipping the download task if the tarball already exists.

Ditching the dict-style configs from bakhti.elk and falling back to a
custom consolidated role for the ELK stack. Still need to port the
Logstash config to Filebeat, but that requires more SSL work.