Configure ElasticSearch snapshot directory via default vars
Force presence of the config line regardless of whether the snapshot
functionality is enabled for the current run of the role. Doing so
ensures that the snapshot functionality will work when required.
Consider flushing handlers conditionally for a more elegant solution.
I suspect we'd need to `wait` after the flush to give ElasticSearch time
to come back up before we request the snapshot.