Adjusts NodeJS args in Kibana service file
Was appending an unecessary "bin/kibana" to the "Environment"
declaration to manage garbage collection for the Node process. Not sure
that it was actually affecting anything, since I'd expect the systemd
parser to discard anything that isn't a key-value pair, but I did notice
the kibana node process hogging up memory. Hopefully this fixes it.