Implements workaround for Elasticsearch indexing
Apparently Elasticsearch 2.0 refuses to index fields with dots in them,
even though the Beat libraries appear to ship logs by default with dots
in the name. It's possible that the dots in the fieldnames only appear
to have dots; for instance, Kibana presents GeoIP fields with dots, but
it's really just collapsing nested fields.