Defers elasticsearch output to end of pipeline
Since the logstash service reads configs in alphabetical order when
building the parsing pipeline, it's important to ensure that the outputs
are handled at the end of the run. Therefore standardizing on:

  * 98 : logstash -> elasticsearch
  * 99 : logstash -> riemann

That order should ensure that all events get stored long-term, and after
all processing has been applied, only then will the events pass to
riemann.