Send alert if Ansible playbook run starts
It'd be nice to be notified when Ansible playbooks are run against our
infrastructure. Tinkering with the throttle filter to identify the start
of the playbook run.

Adds "ansible_command" tag for individual events, and "ansible_playbook"
tag for the start in a series. Not sure if the throttling is working
yet.