Updates target index 'topbeat-*' -> 'logstash-*'
One-line change, but substantial implications. The topbeat logshipper
expects to pass data directly to elasticsearch, rather than logstash,
which was the norm for the deprecated logstash-forwarder. Since this
role munges logs via logstash prior to handing off to elasticsearch for
ingestion and indexing, we want to target the 'logstash-*' index
pattern, because that's where the topbeat event entries land.