Idempotent logfile creation for Kibana
In a classic Ansible two-step, we'll create an empty logfile if no file
exists. If the file does exist, however, Ansible won't force the
permissions, so we'll come back at it in a follow-up task and ensure the
permissions are correct.

Both tasks are skipped conditionally if the logfile for kibana is set to
"stdout".