Activates GeoIP for Redmine logs
Already using GeoIP filters for a variety of data sources, including SSH
and Apache, but the Redmine logs are stored separately, and so require a
distinct filter.

It may make sense to have a catch-all "if [src_ip]" block late in the
filter chain and applies to all log types.