Preserves original log message for SSH alerts
We're overwriting the "message" field for use in notifications, but
that's a destructive action. Prior to overwriting, let's store the
original "message" field (generally the entire raw logline) in a
separate field, called "original_message". This will greatly help with
debugging, and in general make investigation in the Kibana interface
more effective.