Make firewall config optional
On by default, but allows the role to defer firewall management to a
separate role, which is a likely situation in staging or prod. The sane
default is to whitelist IPs, since with minimum configuration the role
should be rather sanely safe.

For the same reason, we're not shipping a default cert with the role,
but rather forcing generation via an external dependency.