Adds URLs for developing logstash filters
The Grok Debugger is well-known and quite useful, if a webapp suits your
fancy for this kind of work. The Grok Constructor I've spent less time
with, but including it because it seems to be a worthwhile resource.

The TDD-based post on testing logstash filters during development was
extremely useful to me. It's still a painful process, but the workflow
proposed in the blog post is feasible, and the least hair-pulling I've
found.