Refines ElasticSearch heap size logic
Per the ES docs [0], the heap size for elasticsearch should be
half of available RAM, but should not exceed 31GB. Setting the var
dynamically via `set_fact`, which prevents overrides. We may want to use
a default var in the future. Also using MB rather than GB to ensure
accurate values with odd numbers of RAM in GB.

[0] https://www.elastic.co/guide/en/elasticsearch/guide/current/heap-sizing.html