Configures pam limits for elasticsearch user
Setting unlimited memlock per the elasticsearch docs (and log messages
for the elasticsearch service). The nofile limit is a fuzzy a number:
significantly higher than the 1000 or so in place on systems where this
elk role is deployed, but more conservative than 100000 or unlimited.

Using defaults for the limits to enable custom overrides.