Catch and alert when Redmine_OpenPGP fails to decrypt message
When the wrong key is used to send email to support@freedom.press, or PGP
ciphertext encrypted to the wrong key is embedded in a contact form submission,
the ticket will be discarded. We should be tagging and alerting on this so that
we can attempt to recover the information.