Parse connection attempts to hidden services on wrong ports
Per OSSEC alerts reported by SecureDrop admins, general crawling of the Tor
network, including with OnionScan, and logs aggregated by FPF, there are
frequent connection attempts to hidden services, on ports other-than-specified.
These result in a warning in the Tor log:

[warn] connection_edge_process_relay_cell (at origin) failed.

This commit adds parsing for an associated info-level logline so we can analyze
these trends and the port numbers that are being tried.