Forces www-data group for htpasswd file
The role was previously reusing the `elk_kibana_user` var for setting
ownership on the htpasswd file for basic auth to the Kibana web UI, but
reusing the var isn't appropriate, since it must be readable by Nginx.
Under Debian, that means www-data by default, so let's hardcode that
value in the task.

Included a comment recognizing the fact that this is Debian-specific,
but so is the entire role at this point.