Handles deprecated symlink for Kibana install destination
The target directory for installing Kibana, /opt/kibana, has not
changed. What has changed, however, is that we now expect that location
to be a directory, whereas it was previously configured as a symlink. To
provide a sane upgrade path to role, let's detect whether it's a
symlink and remove it if so.

Otherwise, existing users of the role will have
multiple versions of Kibana installed into the path, since installing
the deb package will honor the symlink and follow it.