Renames var elk_nginx_user -> elk_kibana_user
The variable is never used anywhere in the Nginx config, so it's no
longer appropriate to call it the nginx user. The service runs as
'kibana' now, and ownership on the log file is set to match this value.
The deb file's postinst hooks have the user hardcoded as 'kibana'.